Allow option to pass in bowerDir explicitly, with default to ./bowerrc.directory as previously defined.
If defined then put all downloaded components into bowerDir, then copy the 'exportsOverride'n files to the targetDir as defined.
If options.targetDir is not defined then default to the derived value of the bowerDir as defined above.  This allows neither value to be set in the task config, allowing both to be defined by the external .bowerrc file; or either or both options can be defined explicitly in task config.